Corrected target name when compiling on 64-bit platforms. Added Callbacks.cpp from dhewm3 - we'll likely kill this down the line and eliminate doomscript entirely in future. Batch for building with vs2022 and x86.

+++ neo/game/script/Script_Interpreter.h
@@ -60,7 +60,7 @@ private:
 
 	void				PopParms( int numParms );
 	void				PushString( const char *string );
-	void				Push( int value );
+	void				Push( intptr_t value );
 	const char			*FloatToString( float value );
 	void				AppendString( idVarDef *def, const char *from );
 	void				SetString( idVarDef *def, const char *from );
@@ -135,12 +135,13 @@ ID_INLINE void idInterpreter::PopParms( int numParms ) {
 idInterpreter::Push
 ====================
 */
-ID_INLINE void idInterpreter::Push( int value ) {
-	if ( localstackUsed + sizeof( int ) > LOCALSTACK_SIZE ) {
+ID_INLINE void idInterpreter::Push( intptr_t value ) {
+	if ( localstackUsed + sizeof( intptr_t ) > LOCALSTACK_SIZE )
+	{
 		Error( "Push: locals stack overflow\n" );
 	}
-	*( int * )&localstack[ localstackUsed ]	= value;
-	localstackUsed += sizeof( int );
+	*( intptr_t * ) &localstack[ localstackUsed ] = value;
+	localstackUsed += sizeof( intptr_t );
 }
 
 /*
